+++ demo/src/main/java/com/filestack/android/demo/MainActivity.java
@@ -9,38 +9,43 @@ import android.util.Log;
 import android.view.View;
 
 import com.filestack.Config;
-import com.filestack.Sources;
-import com.filestack.StorageOptions;
 import com.filestack.android.FsActivity;
 import com.filestack.android.FsConstants;
 import com.filestack.android.Selection;
 
-import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.Locale;
 
 public class MainActivity extends AppCompatActivity {
     private static final int REQUEST_FILESTACK = RESULT_FIRST_USER;
+    private static final String TAG = "MainActivity";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        IntentFilter intentFilter = new IntentFilter(FsConstants.BROADCAST_UPLOAD);
-        UploadStatusReceiver receiver = new UploadStatusReceiver();
-        LocalBroadcastManager.getInstance(this).registerReceiver(receiver, intentFilter);
+        if (savedInstanceState == null) {
+            IntentFilter intentFilter = new IntentFilter(FsConstants.BROADCAST_UPLOAD);
+            UploadStatusReceiver receiver = new UploadStatusReceiver();
+            LocalBroadcastManager.getInstance(this).registerReceiver(receiver, intentFilter);
+        }
     }
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        Locale locale = Locale.getDefault();
+
         super.onActivityResult(requestCode, resultCode, data);
 
         if (requestCode == REQUEST_FILESTACK && resultCode == RESULT_OK) {
-            Log.d("result", "returned");
-            Serializable extra = data.getSerializableExtra(FsConstants.EXTRA_SELECTION_LIST);
-            ArrayList<Selection> selections = (ArrayList<Selection>) extra;
-            for (Selection selection : selections) {
-                Log.i("filestackSelection", selection.getProvider() + " " + selection.getName());
+            Log.i(TAG, "received filestack selections");
+            String key = FsConstants.EXTRA_SELECTION_LIST;
+            ArrayList<Selection> selections = data.getParcelableArrayListExtra(key);
+            for (int i = 0; i < selections.size(); i++) {
+                Selection selection = selections.get(i);
+                String msg = String.format(locale, "selection %d: %s", i, selection.getName());
+                Log.i(TAG, msg);
             }
         }
     }
@@ -52,16 +57,8 @@ public class MainActivity extends AppCompatActivity {
                 getString(R.string.return_url),
                 getString(R.string.policy),
                 getString(R.string.signature));
-        StorageOptions storeOpts = new StorageOptions();
         intent.putExtra(FsConstants.EXTRA_CONFIG, config);
-        intent.putExtra(FsConstants.EXTRA_STORE_OPTS, storeOpts);
         intent.putExtra(FsConstants.EXTRA_AUTO_UPLOAD, true);
-        ArrayList<String> sources = new ArrayList<>();
-        sources.add(Sources.CAMERA);
-        sources.add(Sources.DEVICE);
-        sources.add(Sources.GOOGLE_DRIVE);
-        sources.add(Sources.GITHUB);
-        intent.putExtra(FsConstants.EXTRA_SOURCES, sources);
         startActivityForResult(intent, REQUEST_FILESTACK);
     }