+++ filestack/src/main/java/com/filestack/android/internal/BackButtonListener.java
@@ -1,0 +1,5 @@
+package com.filestack.android.internal;
+
+public interface BackButtonListener {
+    boolean onBackPressed();
+}