+++ filestack/src/main/java/com/filestack/android/internal/Constants.java
@@ -1,0 +1,9 @@
+package com.filestack.android.internal;
+
+import static android.app.Activity.RESULT_FIRST_USER;
+
+public class Constants {
+    public static final int REQUEST_MEDIA_CAPTURE = RESULT_FIRST_USER;
+    // public static final int REQUEST_FILE_BROWSER = RESULT_FIRST_USER + 1;
+    public static final int REQUEST_GALLERY = RESULT_FIRST_USER + 2;
+}