+++ filestack/src/main/java/com/filestack/android/internal/SelectionSaver.java
@@ -1,0 +1,17 @@
+package com.filestack.android.internal;
+
+import com.filestack.android.Selection;
+
+import java.util.ArrayList;
+
+public interface SelectionSaver {
+    boolean toggleItem(Selection selection);
+    boolean isSelected(Selection selection);
+    void setItemChangeListener(Listener listener);
+    ArrayList<Selection> getItems();
+    void clear();
+    boolean isEmpty();
+    interface Listener {
+        void onEmptyChanged(boolean isEmpty);
+    }
+}